updated domain projects

+++ Delivery.BaseLib/Domain/Repositories/IBaseRepository.cs
@@ -8,7 +8,7 @@ public interface IBaseRepository<T> where T : class
     void Add(IEnumerable<T> entities);
     void Update(T entity);
     void Remove(T entity);
-    T? Find(int id);
+    T? Find(long id);
     T? Find(Expression<Func<T, bool>> expression);
     IQueryable<T> Query();
 
@@ -16,7 +16,7 @@ public interface IBaseRepository<T> where T : class
     Task AddAsync(IEnumerable<T> entities, CancellationToken ct);
     Task UpdateAsync(T entity, CancellationToken ct);
     Task RemoveAsync(T entity, CancellationToken ct);
-    Task<T?> FindAsync(int id, CancellationToken ct);
+    Task<T?> FindAsync(long id, CancellationToken ct);
     Task<T?> FindAsync(Expression<Func<T, bool>> expression, CancellationToken ct);
     Task<IQueryable<T>> QueryAsync(CancellationToken ct);
 }
